refactor: file structure for difftester
