+++ src/main/java/com/github/protocolfuzzing/protocolstatefuzzer/statefuzzer/difftester/config/DiffTesterConfig.java
@@ -1,0 +1,5 @@
+package com.github.protocolfuzzing.protocolstatefuzzer.statefuzzer.difftester.config;
+
+public interface DiffTesterConfig {
+
+}
